+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Changelog
 
+- 2026-04-03 MonthRangePicker: Fix popover alignment near right screen edge
 - 2026-03-19 ContextMenu: Add `success` variant